requirements.txt
# Core YOLOv11
ultralytics>=8.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Data Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.10.0

# Monitoring & Visualization
tensorboard>=2.14.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
pandas>=2.0.0

# Development
jupyter>=1.0.0
ipython>=8.15.0
